agent: Reap child processes during the pod lifecycle
All processes, whatever if they come from a new container or to new
processes started by an exec command, are children of cc-agent daemon
running inside the VM. This means that all of them have to be reap
by the parent process after they have terminated.

This patch defines the parent process as the new subreaper and it
starts a go routine aiming to reap any child process after it has
terminated. It does rely on the SIGCHLD signal to determine when
it can expect to reap one of his children.

Fixes #71

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>